+++ aws_emr_launch/constructs/emr_constructs/profile_components.py
@@ -17,7 +17,6 @@ from aws_cdk import (
     aws_kms as kms,
     aws_ec2 as ec2,
     aws_emr as emr,
-    aws_sns as sns,
     core
 )
 
@@ -25,7 +24,7 @@ from ..security_groups.emr import EMRSecurityGroups
 from ..iam_roles.emr import EMRRoles
 
 
-class TransientEMRComponents(core.Construct):
+class EMRProfileComponents(core.Construct):
 
     def __init__(self, scope: core.Construct, id: str, *,
                  cluster_name: str, environment: str, vpc: ec2.Vpc,
@@ -43,15 +42,6 @@ class TransientEMRComponents(core.Construct):
         self._roles = EMRRoles(self, 'TransientEMRComponents_Roles',
                                role_name_prefix='{}-{}'.format(cluster_name, environment),
                                artifacts_bucket=artifacts_bucket, logs_bucket=logs_bucket)
-
-        self._success_topic = sns.Topic(
-            self, 'TransientEMRComponents_SuccessTopic',
-            display_name='{}-{}-SuccessTopic'.format(cluster_name, environment),
-            topic_name='{}-{}-SuccessTopic'. format(cluster_name, environment))
-        self._failure_topic = sns.Topic(
-            self, 'TransientEMRComponents_FailureTopic',
-            display_name='{}-{}-FailureTopic'.format(cluster_name, environment),
-            topic_name='{}-{}-FailureTopic'. format(cluster_name, environment))
 
         self._s3_encryption_mode = None
         self._s3_encryption_key = None
@@ -146,14 +136,6 @@ class TransientEMRComponents(core.Construct):
     @property
     def roles(self) -> EMRRoles:
         return self._roles
-
-    @property
-    def success_topic(self) -> sns.Topic:
-        return self._success_topic
-
-    @property
-    def failure_topic(self) -> sns.Topic:
-        return self._failure_topic
 
     @property
     def s3_encryption_mode(self) -> str: