+++ aws_emr_launch/constructs/iam_roles/__init__.py
@@ -11,3 +11,10 @@
 # express or implied. See the License for the specific language governing
 # permissions and limitations under the License.
 
+from .emr import (
+    EMRRoles,
+    EMREC2InstanceRole,
+    EMRServiceRole,
+    EMRAutoScalingRole
+)
+