+++ aws_emr_launch/lambda_sources/emr_functions/__init__.py
@@ -11,6 +11,7 @@
 # express or implied. See the License for the specific language governing
 # permissions and limitations under the License.
 
+
 def str2bool(v):
     return v.lower() in ("yes", "true", "t", "1")