+++ tests/constructs/test_lambdas.py
@@ -15,7 +15,7 @@ from aws_cdk import (
     core
 )
 
-from aws_emr_launch.constructs.lambdas.emr import EMRLambdas
+from aws_emr_launch.constructs.lambdas import EMRLambdas
 
 
 def test_emr_lambdas():