+++ CHANGELOG.md
@@ -4,17 +4,27 @@ Changelog for emr-launch
 0.1.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- add cluster configuration construct
+
+- refactor and rename components
 
 
 0.1.0 (2019-10-01)
 ------------------
 - emr_components.transient: Security Groups
+
 - emr_components.transient: Roles
+
 - emr_components.transient: Topics
+
 - emr_components.transient: Security Configuration
+
 - iam_roles.emr: Standard Roles
+
 - security_groups.emr: Standard Groups
+
 - lambdas.emr: Add Job Flow
+
 - lambdas.emr: Check Status
+
 - lambdas.emr: Add Step